+++ aptos-move/e2e-tests/src/executor.rs
@@ -930,6 +930,7 @@ impl<O: OutputLogger> FakeExecutorImpl<O> {
         &self,
         txn_block: Vec<Transaction>,
         state_view: &(impl StateView + Sync),
+        check_signature: bool,
     ) -> Result<Vec<TransactionOutput>, VMStatus> {
         // Note: When executing with TransactionSliceMetadata::Block (e.g., in fuzzing/shared-cache
         // modes), the block executor may append a synthetic BlockEpilogue at the end of the block.
@@ -950,7 +951,16 @@ impl<O: OutputLogger> FakeExecutorImpl<O> {
         }
         */
 
-        let sig_verified_block = into_signature_verified_block(txn_block);
+        let sig_verified_block = if check_signature {
+            into_signature_verified_block(txn_block)
+        } else {
+            // This path is only used for comparison testing where txn signature check is not needed
+            let mut verified = vec![];
+            for txn in txn_block {
+                verified.push(SignatureVerifiedTransaction::Valid(txn))
+            }
+            verified
+        };
 
         let mode = self.executor_mode.unwrap_or_else(|| {
             if env::var(ENV_ENABLE_PARALLEL).is_ok() {
@@ -1078,7 +1088,7 @@ impl<O: OutputLogger> FakeExecutorImpl<O> {
         &self,
         txn_block: Vec<Transaction>,
     ) -> Result<Vec<TransactionOutput>, VMStatus> {
-        self.execute_transaction_block_with_state_view(txn_block, &self.state_store)
+        self.execute_transaction_block_with_state_view(txn_block, &self.state_store, true)
     }
 
     pub fn execute_transaction(&self, txn: SignedTransaction) -> TransactionOutput {