+++ crates/aptos/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the Aptos CLI will be captured in this file. This project
 
 # Unreleased
 
+## [7.12.1]
+- Fix link time optimization issues with Aptos CLI released via homebrew.
+
+## [7.12.0]
+- Add support for passing signed integer transaction arguments through the Aptos CLI.
+
 ## [7.11.1]
 - When Move unit tests detect that the legacy struct-based Option module is being loaded, replace it with the pre-compiled binary version.